#7/route classinin tanimlanmasi

+++ app/src/main/java/com/squadro/touricity/message/types/Entry.java
@@ -10,8 +10,8 @@ import lombok.Setter;
 public class Entry {
 
     private String entry_id;
-    private float expense;
-    private float duration;
+    private double expense;
+    private double duration;
     private String comment;
 
     public Entry(String entry_id, float expense, float duration, String comment) {
